Add check of userlimit to JOIN, disable DEBUG otput for parser, fix error in MODE

+++ Include/Channel.hpp
@@ -6,7 +6,7 @@
 /*   By: psenko <psenko@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/22 19:17:28 by psenko            #+#    #+#             */
-/*   Updated: 2025/07/29 16:01:47 by psenko           ###   ########.fr       */
+/*   Updated: 2025/07/30 10:38:27 by psenko           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ class Channel
 		const std::vector<User*> getInvitedUsers(void) const;
 		const std::vector<std::string> getNickList();
 		const std::string getNickListStr();
-		int getUsersCount();
+		size_t getUsersCount();
 		size_t getUserLimit() const;
 		std::string getMode();
